+++ lib/my/my_display_matrices.c
@@ -10,16 +10,16 @@
 
 /* Functions mainly used for debugging */
 
-void disp_integer(void **origin_matrix, unsigned int col, unsigned int row)
+void disp_integer(void **origin_matrix, size_t col, size_t row)
 {
 	int **matrix = (int **)origin_matrix;
 
 	if (!matrix)
 		return;
-	for (unsigned int i = 0; i < col; i++) {
+	for (size_t i = 0; i < col; i++) {
 		if (!matrix[i])
 			return;
-		for (unsigned int j = 0; j < row; j++) {
+		for (size_t j = 0; j < row; j++) {
 			my_put_nbr(matrix[i][j]);
 			my_putchar(' ');
 		}
@@ -27,46 +27,46 @@ void disp_integer(void **origin_matrix, unsigned int col, unsigned int row)
 	}
 }
 
-void disp_double(void **origin_matrix, unsigned int col, unsigned int row)
+void disp_double(void **origin_matrix, size_t col, size_t row)
 {
 	double **matrix = (double **)origin_matrix;
 
 	if (!matrix)
 		return;
-	for (unsigned int i = 0; i < col; i++) {
+	for (size_t i = 0; i < col; i++) {
 		if (!matrix[i])
 			return;
-		for (unsigned int j = 0; j < row; j++)
+		for (size_t j = 0; j < row; j++)
 			printf("%e ", matrix[i][j]);
 		printf("\n");
 	}
 }
 
-void disp_float(void **origin_matrix, unsigned int col, unsigned int row)
+void disp_float(void **origin_matrix, size_t col, size_t row)
 {
 	float **matrix = (float **)origin_matrix;
 
 	if (!matrix)
 		return;
-	for (unsigned int i = 0; i < col; i++) {
+	for (size_t i = 0; i < col; i++) {
 		if (!matrix[i])
 			return;
-		for (unsigned int j = 0; j < row; j++)
+		for (size_t j = 0; j < row; j++)
 			printf("%f ", matrix[i][j]);
 		printf("\n");
 	}
 }
 
-void disp_char(void **origin_matrix, unsigned int col, unsigned int row)
+void disp_char(void **origin_matrix, size_t col, size_t row)
 {
 	char **matrix = (char **)origin_matrix;
 
 	if (!matrix)
 		return;
-	for (unsigned int i = 0; i < col; i++) {
+	for (size_t i = 0; i < col; i++) {
 		if (!matrix[i])
 			return;
-		for (unsigned int j = 0; j < row; j++) {
+		for (size_t j = 0; j < row; j++) {
 			my_putchar(matrix[i][j]);
 			my_putchar(' ');
 		}
@@ -74,17 +74,17 @@ void disp_char(void **origin_matrix, unsigned int col, unsigned int row)
 	}
 }
 
-void display_matrix(void **matrix, unsigned int col,
-	unsigned int row, char type)
+void display_matrix(void **matrix, size_t col,
+	size_t row, char type)
 {
 	char types[5] = "idfc";
-	void (*display_method[4])(void **, unsigned int, unsigned int) =
+	void (*display_method[4])(void **, size_t, size_t) =
 		{disp_integer,
 		disp_double,
 		disp_float,
 		disp_char};
 
-	for (unsigned int i = 0; i < 4; i++)
+	for (size_t i = 0; i < 4; i++)
 		if (types[i] == type)
 			(*display_method[i])(matrix, col, row);
 }
