+++ lib/my/my_strnrevcmp.c
@@ -7,14 +7,14 @@
 
 #include "my.h"
 
-int my_strnrevcmp(const char *s1, const char *s2, unsigned int n)
+int my_strnrevcmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int count = my_strlen(s1);
-	unsigned int count2 = my_strlen(s2);
+	size_t count = my_strlen(s1);
+	size_t count2 = my_strlen(s2);
 
 	if (!s1 || !s2)
 		return (-1);
-	for (unsigned int i = 0; s1[count] == s2[count2] && count != 0
+	for (size_t i = 0; s1[count] == s2[count2] && count != 0
 	&& count2 != 0 && i < n; i++) {
 		count--;
 		count2--;
